[breaking] Moving all spectra to hazma.spectra
As things are now, we have 'decay', 'neutrino_spectra'
and 'positron_spectra' modules. To make things clearer,
I am moving all of these to the 'spectra' module. I've
also given all functions a consistant names of
'dnde_f_p' where f is 'photon', 'positron' or 'neutrino'
and p is the name of the decaying particle.